Fix #201

+++ arm9/source/fileOperations.cpp
@@ -15,7 +15,7 @@
 #define copyBufSize 0x8000
 #define shaChunkSize 0x10000
 
-u32* copyBuf = (u32*)0x02004000;
+u8* copyBuf = (u8*)0x02004000;
 
 std::vector<ClipboardFile> clipboard;
 bool clipboardOn = false;
